carga_horaria integer

+++ migrations/m200306_002031_Criar_Tabela_Disciplinas_Disponiveis.php
@@ -16,7 +16,7 @@ class m200306_002031_Criar_Tabela_Disciplinas_Disponiveis extends Migration
             [
                 'id'=>$this->primaryKey(),
                 'descricao'=>$this->string(60)->notNull(),
-                'carga_horaria'=>$this->float()->notNull(),
+                'carga_horaria'=>$this->integer()->notNull(),
             ]);
     }
